requirements.txt
langchain-core
langchain-community
langchain-openai
langchain-chroma
langchain-experimental
unstructured[all-docs]
pillow
pydantic
lxml
pillow
matplotlib
chromadb
tiktoken
python-dotenv
openai
streamlit
python-magic-bin
ipykernel